requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.0
fastapi-cli==0.0.5
filelock==3.15.4
fsspec==2024.6.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.7.24
rich==13.7.1
scikit-learn==1.5.1
scipy==1.14.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.1
textblob==0.18.0.post0
threadpoolctl==3.5.0
torch==2.2.2
tqdm==4.66.5
typer==0.12.3
typing_extensions==4.12.2
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.23.0
websockets==12.0
